x86: lib: zimage: Add support for EFI applications
When U-Boot is built as an EFI application, it needs to pass the EFI
information to the Linux kernel in a different way than when built as an
EFI stub. This patch adds support for setting up the EFI information
in the boot_params structure when U-Boot is built as an EFI application.

Signed-off-by: Matheus Castello <matheus@castello.eng.br>